chore(selection): refactor spec
* move SelectionBehavior tests to SelectionBehaviorSpec
* update comments
