+++ lib/features/selection/SelectionBehavior.js
@@ -1,29 +1,28 @@
 import {
-  hasPrimaryModifier, hasSecondaryModifier
+  hasPrimaryModifier,
+  hasSecondaryModifier
 } from '../../util/Mouse';
 
 import {
-  find, isArray
+  find,
+  isArray
 } from 'min-dash';
 
 
-export default function SelectionBehavior(
-    eventBus, selection, canvas,
-    elementRegistry) {
+export default function SelectionBehavior(eventBus, selection, canvas, elementRegistry) {
 
-  eventBus.on('create.end', 500, function(e) {
-
-    var context = e.context,
+  // Select elements on create
+  eventBus.on('create.end', 500, function(event) {
+    var context = event.context,
         canExecute = context.canExecute,
         elements = context.elements,
         hints = context.hints || {},
         autoSelect = hints.autoSelect;
 
-    // select elements after they have been created
     if (canExecute) {
       if (autoSelect === false) {
 
-        // select no elements
+        // Select no elements
         return;
       }
 
@@ -31,44 +30,43 @@ export default function SelectionBehavior(
         selection.select(autoSelect);
       } else {
 
-        // select all elements by default
+        // Select all elements by default
         selection.select(elements.filter(isShown));
       }
     }
   });
 
-  eventBus.on('connect.end', 500, function(e) {
+  // Select connection targets on connect
+  eventBus.on('connect.end', 500, function(event) {
+    var context = event.context,
+        canExecute = context.canExecute,
+        hover = context.hover;
 
-    // select the connect end target
-    // after a connect operation
-    if (e.context.canExecute && e.context.hover) {
-      selection.select(e.context.hover);
+    if (canExecute && hover) {
+      selection.select(hover);
     }
   });
 
-  eventBus.on('shape.move.end', 500, function(e) {
-    var previousSelection = e.previousSelection || [];
+  // Select shapes on move
+  eventBus.on('shape.move.end', 500, function(event) {
+    var previousSelection = event.previousSelection || [];
 
-    var shape = elementRegistry.get(e.context.shape.id);
+    var shape = elementRegistry.get(event.context.shape.id);
 
-    // make sure at least the main moved element is being
-    // selected after a move operation
-    var inSelection = find(previousSelection, function(selectedShape) {
+    // Always select main shape on move
+    var isSelected = find(previousSelection, function(selectedShape) {
       return shape.id === selectedShape.id;
     });
 
-    if (!inSelection) {
+    if (!isSelected) {
       selection.select(shape);
     }
   });
 
-  // Shift + click selection
+  // Select elements on click
   eventBus.on('element.click', function(event) {
-
     var element = event.element;
 
-    // do not select the root element
-    // or connections
     if (element === canvas.getRootElement()) {
       element = null;
     }
@@ -76,20 +74,26 @@ export default function SelectionBehavior(
     var isSelected = selection.isSelected(element),
         isMultiSelect = selection.get().length > 1;
 
-    // mouse-event: SELECTION_KEY
+    // Add to selection if CTRL or SHIFT pressed
     var add = hasPrimaryModifier(event) || hasSecondaryModifier(event);
 
-    // select OR deselect element in multi selection
     if (isSelected && isMultiSelect) {
       if (add) {
+
+        // Deselect element
         return selection.deselect(element);
       } else {
+
+        // Select element only
         return selection.select(element);
       }
-    } else
-    if (!isSelected) {
+    } else if (!isSelected) {
+
+      // Select element
       selection.select(element, add);
     } else {
+
+      // Deselect element
       selection.deselect(element);
     }
   });
